+++ public/components/all/package.go
@@ -8,6 +8,7 @@ import (
 	_ "github.com/Jeffail/benthos/v3/public/components/legacy"
 
 	// Import new service packages.
+	_ "github.com/Jeffail/benthos/v3/internal/impl/aws"
 	_ "github.com/Jeffail/benthos/v3/internal/impl/confluent"
 	_ "github.com/Jeffail/benthos/v3/internal/impl/gcp"
 	_ "github.com/Jeffail/benthos/v3/internal/impl/mongodb"